Move some some from paymentRepository --> paymentUtils, add linkUtils
