+++ packages/hypernet-core/src/implementations/business/LinkService.ts
@@ -6,6 +6,9 @@ import { HypernetLink } from "@interfaces/objects";
 export class LinkService implements ILinkService {
     constructor(protected ledgerRepository: ILinkRepository) {}
     
+    /**
+     * 
+     */
     public async getLedgers(): Promise<HypernetLink[]> {
         return this.ledgerRepository.getHypernetLinks();
     }
