+++ packages/hypernet-core/src/implementations/business/PaymentService.ts
@@ -2,15 +2,11 @@ import { IPaymentService } from "@interfaces/business";
 import { IAccountsRepository, ILinkRepository } from "@interfaces/data";
 import { IPaymentRepository } from "@interfaces/data/IPaymentRepository";
 import {
-    HypernetLink,
     BigNumber,
     Payment,
     EthereumAddress,
     PublicKey,
-    ControlClaim,
-    HypernetConfig,
     PublicIdentifier,
-    Balances,
     PushPayment,
     PullPayment
   } from "@interfaces/objects";
@@ -21,13 +17,12 @@ import { IConfigProvider, IContextProvider } from "@interfaces/utilities";
  * PaymentService uses Vector internally to send payments on the requested channel.
  */
 export class PaymentService implements IPaymentService {
-    constructor(protected linkRepository: ILinkRepository,
+    constructor(
+        protected linkRepository: ILinkRepository,
         protected accountRepository: IAccountsRepository,
         protected contextProvider: IContextProvider,
         protected configProvider: IConfigProvider,
-        protected paymentRepository: IPaymentRepository) {
-
-    }
+        protected paymentRepository: IPaymentRepository) {}
 
     /**
      * 
@@ -36,7 +31,7 @@ export class PaymentService implements IPaymentService {
     public async acceptFunds(paymentIds: string[]): Promise<Payment[]> {
         // Get the payments from the repo, to make sure it can be accepted.
         const config = await this.configProvider.getConfig();
-        const payments = await this.paymentRepository.getPaymentsById(paymentIds);
+        const payments = await this.paymentRepository.getPaymentsByIds(paymentIds);
         const hypertokenBalance = await this.accountRepository.getBalanceByAsset(config.hypertokenAddress)
 
         // Loop over the payments and do a sanity check
@@ -68,7 +63,7 @@ export class PaymentService implements IPaymentService {
      * Called by the reciever of a parameterized transfer
      */
     public async paymentPosted(paymentId: string): Promise<void> {
-        const payments = await this.paymentRepository.getPaymentsById([paymentId]);
+        const payments = await this.paymentRepository.getPaymentsByIds([paymentId]);
         const payment = payments.get(paymentId);
 
         if (payment == null || payment.state != EPaymentState.Approved) {
@@ -87,15 +82,23 @@ export class PaymentService implements IPaymentService {
         }
     }
 
+    /**
+     * 
+     * @param paymentId 
+     */
     public async pullRecorded(paymentId: string): Promise<void> {
-        const payments = await this.paymentRepository.getPaymentsById([paymentId]);
+        const payments = await this.paymentRepository.getPaymentsByIds([paymentId]);
         const payment = payments.get(paymentId);
         
         throw new Error('Method not yet implemented')
     }
 
+    /**
+     * 
+     * @param paymentId 
+     */
     public async stakePosted(paymentId: string): Promise<void> {
-        const payments = await this.paymentRepository.getPaymentsById([paymentId]);
+        const payments = await this.paymentRepository.getPaymentsByIds([paymentId]);
         const payment = payments.get(paymentId);
     
         if (payment == null || payment.state != EPaymentState.Staked) {
@@ -115,7 +118,7 @@ export class PaymentService implements IPaymentService {
      */
     public async offerReceived(paymentId: string): Promise<void> {
         
-        const paymentsPromise =  this.paymentRepository.getPaymentsById([paymentId]);
+        const paymentsPromise =  this.paymentRepository.getPaymentsByIds([paymentId]);
         const contextPromise = this.contextProvider.getInitializedContext();
         const [payments, context] = await Promise.all([paymentsPromise, contextPromise])
