+++ packages/hypernet-core/src/implementations/utilities/VectorUtils.ts
@@ -1,19 +1,17 @@
-import { BrowserNode } from "@connext/vector-browser-node";
 import { FullChannelState, FullTransferState, PublicIdentifier } from "@connext/vector-types";
 import { IHypernetTransferMetadata } from "@interfaces/objects";
-import { ETransferType } from "@interfaces/types";
-import { IBrowserNodeProvider, IContextProvider, IVectorUtils, IConfigProvider, ISortedTransfers } from "@interfaces/utilities";
+import { IBrowserNodeProvider, IContextProvider, IVectorUtils, IConfigProvider } from "@interfaces/utilities";
 
 export class VectorUtils implements IVectorUtils {
     protected channelAddress: string | null;
 
-    constructor(protected configProvider: IConfigProvider,
+    constructor(
+        protected configProvider: IConfigProvider,
         protected contextProvider: IContextProvider,
         protected browserNodeProvider: IBrowserNodeProvider) { 
 
         this.channelAddress = null;
     }
-
 
     public async createNullTransfer(counterParty: PublicIdentifier, metadata: IHypernetTransferMetadata): Promise<FullTransferState> {
         throw new Error('Method not yet implemented')
@@ -94,86 +92,5 @@ export class VectorUtils implements IVectorUtils {
         this.channelAddress = newChannel.channelAddress;
         return this.channelAddress
     }
-
-    public getTransferType(transfer: FullTransferState): ETransferType {
-        // Have to jump through some hoops here
-        // TODO: Fix this.
-        return ETransferType.Offer;
-    }
-
-    /**
-     * 
-     * @param paymentId 
-     * @param transfers 
-     * @param browserNode 
-     */
-    public async sortTransfers(
-        paymentId: string,
-        transfers: FullTransferState[],
-        browserNode: BrowserNode): Promise<SortedTransfers> {
-
-        // We need to do a lookup for non-active transfers for the payment ID.
-        // TODO
-        // let inactiveTransfers = await browserNode.getTransfers((transfer) => {return transfer.meta.paymentId == fullPaymentId;});
-        // transfers.concat(inactiveTransfers);
-
-        const offerTransfers = transfers.filter((val) => {
-            return this.getTransferType(val) == ETransferType.Offer;
-        });
-        
-        const insuranceTransfers = transfers.filter((val) => {
-            return this.getTransferType(val) == ETransferType.Insurance;
-        });
-        
-        const parameterizedTransfers = transfers.filter((val) => {
-            return this.getTransferType(val) == ETransferType.Parameterized;
-        });
-
-        const pullTransfers = transfers.filter((val) => {
-            return this.getTransferType(val) == ETransferType.PullRecord;
-        });
-
-        const unrecognizedTransfers = transfers.filter((val) => {
-            return this.getTransferType(val) == ETransferType.Unrecognized;
-        });
-
-        if (unrecognizedTransfers.length > 0) {
-            throw new Error("Payment includes unrecognized transfer types!")
-        }
-
-        if (offerTransfers.length != 1) {
-            // TODO: this could be handled more elegantly; if there's other payments
-            // but no offer, it's still a valid payment
-            throw new Error("Invalid payment, no offer transfer!")
-        }
-        const offerTransfer = offerTransfers[0];
-
-        let insuranceTransfer: FullTransferState | null = null;
-        if (insuranceTransfers.length == 1) {
-            insuranceTransfer = insuranceTransfers[0];
-        } else if (insuranceTransfers.length > 1) {
-            throw new Error("Invalid payment, too many insurance transfers!");
-        }
-
-        let parameterizedTransfer: FullTransferState | null = null;
-        if (parameterizedTransfers.length == 1) {
-            parameterizedTransfer = parameterizedTransfers[0];
-        } else if (parameterizedTransfers.length > 1) {
-            throw new Error("Invalid payment, too many parameterized transfers!");
-        }
-
-        return new SortedTransfers(offerTransfer,
-            insuranceTransfer,
-            parameterizedTransfer,
-            pullTransfers,
-            offerTransfer.meta);
-    }
 }
 
-export class SortedTransfers implements ISortedTransfers {
-    constructor(public offerTransfer: FullTransferState,
-        public insuranceTransfer: FullTransferState | null,
-        public parameterizedTransfer: FullTransferState | null,
-        public pullRecordTransfers: FullTransferState[],
-        public metadata: IHypernetTransferMetadata) { }
-}
