+++ packages/hypernet-core/src/interfaces/business/ILinkService.ts
@@ -1,5 +1,9 @@
 import { HypernetLink } from "@interfaces/objects";
 
 export interface ILinkService {
+
+    /**
+     * 
+     */
     getLedgers(): Promise<HypernetLink[]>;
 }
