+++ packages/hypernet-core/src/interfaces/utilities/ILinkUtils.ts
@@ -1,0 +1,6 @@
+import { HypernetLink, InitializedHypernetContext, Payment } from "@interfaces/objects";
+
+export interface ILinkUtils {
+    paymentsToHypernetLinks(payments: Payment[], context: InitializedHypernetContext): Promise<HypernetLink[]> 
+    
+}
